잘못 머지해서 다시 커밋.

+++ todo-app/src/components/ToDoTemplate.js
@@ -22,6 +22,10 @@ const ToDoTemplate = () => {
 
   const onInsert = useCallback(
     async (subject) => {
+      if (subject.trim().length === 0) {
+        alert('할 일을 입력해주세요.');
+        return;
+      }
       try {
         await axios({
           method: 'post',
@@ -40,6 +44,7 @@ const ToDoTemplate = () => {
 
   const onRemove = useCallback(
     async (id) => {
+      if (!validCheckId(id)) return;
       try {
         await axios.delete(`http://localhost:3001/todos/${id}`);
         await fetchData();
@@ -52,6 +57,7 @@ const ToDoTemplate = () => {
 
   const onToggle = useCallback(
     async (id) => {
+      if (!validCheckId(id)) return;
       const todo = todos.find((todo) => todo.id === id);
       if (!todo) {
         return;
@@ -67,6 +73,14 @@ const ToDoTemplate = () => {
     },
     [todos],
   );
+
+  const validCheckId = (id) => {
+    if (todos.findIndex((todo) => todo.id === id) === -1) {
+      alert('잘못된 id입니다.');
+      return false;
+    }
+    return true;
+  };
 
   return (
     <div className="TodoTemplate">
